Dispersal

Dispersal.jl

Dispersal — Module.

This package extends Cellular.jl

Dispersal.jl provides a range of dispersal modules that can be combined to build grid-based organism dispersal simulations.

The design provides a solid framework while allowing customisation of any aspect. A model may start with the defaults and formulations provided, but incrementally customise them for a particular use-case, to any level of detail.

Additionally, modules, outputs, neighborhoods provided by Cellular.jl or other packages that extend it may be incorporated into a simulaiton.

source

Dispersal.jl

Example

Models

Neighborhoods

Layers

Example

This is the general pattern of running model in Dispersal.jl.

We declare a suitability layer and initial conditions. Then we run a dispersal simulation that combines local and jump dispersal, for three timesteps.

using Cellular
using Dispersal

# add a 0.0-1.0 scaled raster array that represents habitat suitability
suitability = [0.5 0.0 0.3 0.0 0.0 0.3 0.0;
               0.0 0.2 0.8 0.0 0.9 0.6 0.0;
               0.0 0.5 1.0 1.0 1.0 0.0 0.0;
               0.0 0.0 1.0 1.0 0.7 0.0 0.0;
               0.0 0.0 0.6 1.0 0.0 0.0 0.0;
               0.0 0.1 0.6 0.0 0.0 0.0 0.0;
               0.0 0.0 0.0 0.0 0.0 0.0 0.7]

# Make an init array the same size as your suitability layer, and seed it
init = [0  0  0  0  0  0  0;
        0  0  0  0  0  0  0;
        0  0  0  0  0  0  0;
        0  0  0  0  0  0  0;
        0  0  1  0  0  0  0;
        0  0  0  0  0  0  0;
        0  0  0  0  0  0  0]

# You also could do this with:
# init = zeros(suitability); init[5, 3] = 1

# Define a dispersal kernel function

f = d -> e^-d

# Define the neighborhood, using the dispersal kernel and a radius
hood = DispersalNeighborhood(; f=f, radius=2)

# Define additional raster layers
layers = SuitabilityLayer(suitability)

# Define disersal modules
localdisp = InwardsLocalDispersal(layers=layers, neighborhood=hood)
jumpdisp = JumpDispersal(layers=layers)

# Set the output type
output = ArrayOutput(init)

# Run the simulation
sim!(output, (localdisp, jumpdisp), init; time=1:3)

output.frames[3]

7×7 Array{Int64,2}:
 0  0  0  0  0  0  0
 0  0  0  0  0  0  0
 0  0  0  0  0  0  0
 0  0  0  0  0  0  0
 0  0  0  0  0  0  0
 0  0  0  0  0  0  0
 0  0  0  0  0  0  0

Models

Extending AbstractModel

Dispersal.AbstractInwardsLocalDispersal — Type.

abstract AbstractInwardsLocalDispersal <: Cellular.AbstractModel

source

Dispersal.InwardsLocalDispersal — Type.

Local dispersal within a DispersalNeighborhood or other neighborhoods. Inwards dispersal calculates dispersal to the current cell from cells in the neighborhood.

source

Extending AbstractPartialModel

These models trigger rules that only partially update the grid. Many dispersal functions only operate on cells that are currently occupied.

Dispersal.AbstractOutwardsLocalDispersal — Type.

abstract AbstractOutwardsLocalDispersal <: Cellular.AbstractPartialModel

source

Dispersal.OutwardsLocalDispersal — Type.

Local dispersal within a DispersalNeighborhood

Outwards dispersal calculates dispersal from the current cell to cells in its neighborhood. This should be more efficient than inwards dispersal when a small number of cells are occupied, but less efficient when a large proportion of the grid is occupied.

source

Dispersal.AbstractJumpDispersal — Type.

abstract AbstractJumpDispersal <: Cellular.AbstractPartialModel

source

Dispersal.JumpDispersal — Type.

Jump dispersal within a DispersalNeighborhood] or other neighborhoods.

source

Dispersal.AbstractHumanDispersal — Type.

abstract AbstractHumanDispersal <: Cellular.AbstractPartialModel

source

Dispersal.HumanDispersal — Type.

Human dispersal model.

source

Models trigger methods of rule. Custom models you create will in most cases need a custom rule method.

Cellular.rule — Function.

rule(model::AbstractInwardsLocalDispersal, state, index, t, args...)

Runs rule for of AbstractInwardsLocalDispersal dispersal.

The current cell is invaded if there is pressure from surrounding cells and suitable habitat. Otherwise it keeps its current state.

source

rule(model::AbstractOutwardsLocalDispersal, state, index, t, source, dest, args...)

Runs rule for of AbstractOutwardsLocalDispersal dispersal.

Surrounding cells are invaded if the current cell is occupied and they have suitable habitat. Otherwise they keeps their current state.

source

rule(model::AbstractJumpDispersal, state, index, t, source, dest, args...)

Long range rule for AbstractJumpDispersal. A random cell within the spotrange is invaded if it is suitable.

source

rule(model::AbstractHumanDispersal, state, index, t, source, dest, args...)

Simulates human dispersal, weighting dispersal probability based on human population in the source cell.

source

Neighborhoods

Extend Cellular.AbstractNeighborhood, and add neighbors() methods.

Types and Constructors

Dispersal.AbstractDispersalNeighborhood — Type.

abstract AbstractDispersalNeighborhood <: Cellular.AbstractNeighborhood

source

Dispersal.DispersalNeighborhood — Type.

struct DispersalNeighborhood{K, S} <: Dispersal.AbstractDispersalNeighborhood

dispkernel

radius

overflow

source

Dispersal.DispersalNeighborhood — Method.

DispersalNeighborhood(; f=d -> exponential(d, 1), radius=3, overflow=Skip())

Constructor for neighborhoods, using a dispersal kernel function and a cell radius.

Keyword Arguments:

f::Function: any function that accepts a Number argument and returns a propbability between 0.0 and 1.0

radius::Integer: a positive integer

source

Methods

neighbors
neighbors(hood::DispersalNeighborhood, state, index, t, source, args...)
pressure

Layers

Layers are a concept not present in Cellular.jl. They provide overlay grids of additional information about dispersal potential.

Like models, than can be combined arbitrarily in tuples. Methods loop through all relevant layers to return a scalar that is the product of their outputs.

Types

Dispersal.AbstractLayer — Type.

abstract AbstractLayer

source

Dispersal.AbstractSuitabilityLayer — Type.

abstract AbstractSuitabilityLayer <: Dispersal.AbstractLayer

source

Dispersal.SuitabilityLayer — Type.

struct SuitabilityLayer{T} <: Dispersal.AbstractSuitabilityLayer

data

Any 2-dimensional AbstractArray matching the coordinates of the init array

source

Dispersal.AbstractSuitabilitySequence — Type.

abstract AbstractSuitabilitySequence <: Dispersal.AbstractSuitabilityLayer

source

Dispersal.SuitabilitySequence — Type.

struct SuitabilitySequence{T, D} <: Dispersal.AbstractSuitabilitySequence

timespan

The timespan of each layer in the sequence.

data

Either an Array of 2-dimensional arrays matching the coordinates of the init array, or a similar 3 dimensional array where the 3rd dimension is the time-step.

source

Dispersal.HumanLayer — Type.

struct HumanLayer{T} <: Dispersal.AbstractHumanLayer

data

Any 2-dimensional AbstractArray matching the coordinates of the init array

source

Methods

Dispersal.suitability — Function.

suitability(layers, row::Int, col::Int, t::Number)

Returns a scalar representing cell suitability from one or multiple layers.

For multiple layers the product of each individual scalar is returned.

Layers of type other than AbstractSuitabilityLayer return 1.0.

Arguments

layers : a single layer or tuple of layers of any type

row::Int

col::Int

t::Number : current timestep for interploating layere sequences.

source

Dispersal.human_impact — Function.

human_impact(layers, row::Int, col::Int, t)

Returns a scalar indicating human impact from one or multiple layers.

For multiple layers the product of each individual scalar is returned.

Layers of type other than AbstractHumanLayer return 1.0.

Arguments

layers : a single layer or tuple of layers of any type

row::Int

col::Int

t::Number : current timestep for interploating layere sequences.

source

Dispersal.sequence_interpolate — Function.

sequence_interpolate(layer, row, col, t)

Interpolates between layers in a sequence.

source

Dispersal.cyclic — Function.

Cycles a time position through a particular timespan length

source