Introduction

Dispersal.jl

Dispersal — Module.

This package extends Cellular.jl

Dispersal.jl provides a range of dispersal modules that can be combined to build grid-based organism dispersal simulations.

The design provides a solid framewhork while allowing customisation of any aspect of a simulation. A model may start with the defaults and formulations provided, but incrementally customise them for a particular use-case, to any level of detail.

Additionally, modules, outputs, neighborhoods provided by Cellular.jl or other packages that extend it may all be incorporated into a simulaiton.

source

Example

This is the general pattern of running model in Dispersal.jl.

We declare a suitability layer and initial conditions. Then we run a dispersal simulation that combines local and jump dispersal, for three timesteps.

using Cellular
using Dispersal

# add a 0.0-1.0 scaled raster array that represents habitat suitability
suitability = [0.5 0.0 0.3 0.0 0.0 0.3 0.0;
               0.0 0.2 0.8 0.0 0.9 0.6 0.0;
               0.0 0.5 1.0 1.0 1.0 0.0 0.0;
               0.0 0.0 1.0 1.0 0.7 0.0 0.0;
               0.0 0.0 0.6 1.0 0.0 0.0 0.0;
               0.0 0.1 0.6 0.0 0.0 0.0 0.0;
               0.0 0.0 0.0 0.0 0.0 0.0 0.7]

# Make an init array the same size as your suitability layer, and seed it
init = [0  0  0  0  0  0  0;
        0  0  0  0  0  0  0;
        0  0  0  0  0  0  0;
        0  0  0  0  0  0  0;
        0  0  1  0  0  0  0;
        0  0  0  0  0  0  0;
        0  0  0  0  0  0  0]

# You also could do this with:
# init = zeros(suitability); init[5, 3] = 1

# Define a dispersal kernel function

f = d -> e^-d

# Define the neighborhood, using the dispersal kernel and a radius
hood = DispersalNeighborhood(; f=f, radius=2)

# Define additional raster layers
layers = SuitabilityLayer(suitability)

# Define disersal modules
localdisp = LocalDispersal(layers=layers, neighborhood=hood)
jumpdisp = JumpDispersal(layers=layers)

# Set the output type
output = ArrayOutput(init)

# Run the simulation
sim!(output, (localdisp, jumpdisp), init; time=1:3)

output.frames[3]

7×7 Array{Int64,2}:
 1  0  1  0  0  1  0
 0  1  1  0  1  1  0
 0  1  1  1  1  0  0
 0  0  1  1  1  0  0
 0  0  1  1  0  0  0
 0  0  1  0  0  0  0
 0  0  0  0  0  0  0

Models

Dispersal.AbstractLocalDispersal — Type.

abstract AbstractLocalDispersal <: Cellular.AbstractModel

Extend to modify LocalDispersal

source

Dispersal.LocalDispersal — Type.

struct LocalDispersal{N, L, P, S} <: Dispersal.AbstractLocalDispersal

Local dispersal within a DispersalNeighborhood] or other AbstractNeighborhood

Keyword Arguments

neighborhood: = DispersalNeighborhood()

prob: A real number between one and zero.

layers: AbstractLayers or a single AbstractLayer. The default is nothing.

cellsize::S = A number or Unitful.jl distance.

source

Dispersal.AbstractJumpDispersal — Type.

abstract AbstractJumpDispersal <: Cellular.AbstractInPlaceModel

Extend to modify JumpDispersal

source

Dispersal.JumpDispersal — Type.

struct JumpDispersal{L, S} <: Dispersal.AbstractJumpDispersal

Local dispersal within a DispersalNeighborhood] or other AbstractNeighborhood

Keyword Arguments

spotrange: A number or Unitful.jl distance.

prob: A real number between one and zero.

layers: AbstractLayers or a single AbstractLayer. The default is nothing.

cellsize::S = A number or Unitful.jl distance.

source

Dispersal.AbstractHumanDispersal — Type.

abstract AbstractHumanDispersal <: Cellular.AbstractInPlaceModel

Inherit to exten human dispersal

source

Dispersal.HumanDispersal — Type.

struct HumanDispersal{L, S} <: Dispersal.AbstractHumanDispersal

Human dispersal model.

human

layers

cellsize

source

These modules all trigger methods of rule:

Cellular.rule — Function.

rule(model, state, index, t, args)

Short range rule for AbstractLocalDispersal dispersal. Cells are invaded if there is pressure and suitable habitat, otherwise left as-is.

source

rule(model, state, index, t, source, args)

Long range rule for AbstractJumpDispersal. A random cell within the spotrange is invaded if it is suitable.

source

Neighborhoods

Extend Cellular.AbstractNeighborhood

Dispersal.AbstractDispersalNeighborhood — Type.

abstract AbstractDispersalNeighborhood <: Cellular.AbstractNeighborhood

Neighborhoods for dispersal

source

Dispersal.DispersalNeighborhood — Type.

struct DispersalNeighborhood{K, S} <: Dispersal.AbstractDispersalNeighborhood

A neighborhood built from a dispersal kernel function and a cell radius.

Arguments:

dispkernel::AbstractArray{T,2}

overflow::AbstractOverflow

source

Customisation

Cellular.neighbors — Function.

neighbors(h, state, index, t, source, args)

Returns nieghbors for a DispersalNeighborhood, looping over the array of dispersal propabilities.

source

Dispersal.build_dispersal_kernel — Function.

build_dispersal_kernel(f, r)

Accepts a dispersal kernel function and integer radius, and returns an array of probabilities, of size 2r + 1 * 2r + 1.

source

Dispersal.pressure — Function.

pressure(model, cc)

defined at Dispersal/src/disperse.jl:122.

Calculates the propagule pressure from the output of a neighborhood.

source

Layers

Layers are a concept not present in Cellular.jl. They provide overlay grids of additional information about dispersal potential. Raw arrays are wrapped in a type that determines how they will be used.

Dispersal.AbstractLayer — Type.

abstract AbstractLayer

Layers wrap an array, normally built from some kind of raster.

The wrapper defines the purpose of the layer and allows specialised method dispatch to utilise them.

source

Dispersal.AbstractSuitabilityLayer — Type.

abstract AbstractSuitabilityLayer <: Dispersal.AbstractLayer

source

Dispersal.SuitabilityLayer — Type.

struct SuitabilityLayer{T} <: Dispersal.AbstractSuitabilityLayer

data

source

Dispersal.AbstractSuitabilitySequence — Type.

abstract AbstractSuitabilitySequence <: Dispersal.AbstractSuitabilityLayer

source

Dispersal.SuitabilitySequence — Type.

struct SuitabilitySequence{T, D} <: Dispersal.AbstractSuitabilitySequence

timespan

data

source

Dispersal.HumanLayer — Type.

struct HumanLayer{T} <: Dispersal.AbstractLayer

data

source

Dispersal.AbstractLayers — Type.

abstract AbstractLayers

source

Dispersal.Layers — Type.

struct Layers{S, F} <: Dispersal.AbstractLayers

suitability

human

source

Customisation

Dispersal.suitability — Function.

suitability(layers, row, col, t)

Returns a suitability scalar from a single layer, or the product of multiple layers

source

Dispersal.seq_interpolate — Function.

seq_interpolate(layer, row, col, t)

Interpolates between layers in a sequence

source

Dispersal.get_cell — Function.

get_cell(layer, row, col, pos)

Return a particular cell from a layer, given row, column and timestep)

source